+++ iblrig/gui/wizard.py
@@ -189,7 +189,10 @@ class RigWizardModel:
                     raise e
 
         # since we are connecting to Alyx, validate some parameters to ensure a smooth extraction
-        result = iblrig.hardware_validation.ValidateAlyxLabLocation().run(self.alyx)
+        result = iblrig.hardware_validation.ValidateAlyxLabLocation(
+            iblrig_settings=self.iblrig_settings,
+            hardware_settings=self.hardware_settings,
+        ).run(self.alyx)
         if result.status == 'FAIL' and gui:
             QtWidgets.QMessageBox().critical(None, 'Error', f'{result.message}\n\n{result.solution}')
 
@@ -774,6 +777,8 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
 
                 case 'session_template_id':
                     label = 'Session Template ID'
+                    widget.setMinimum(0)
+                    widget.setMaximum(11)
 
                 case 'delay_secs':
                     label = 'Initial Delay, s'
@@ -1015,7 +1020,11 @@ class RigWizard(QtWidgets.QMainWindow, Ui_wizard):
                 session_data = json.load(fid)
 
             # check if session was a dud
-            if (ntrials := session_data['NTRIALS']) < 42 and 'spontaneous' not in self.model.task_name:
+            if (
+                (ntrials := session_data['NTRIALS']) < 42
+                and not any([x in self.model.task_name for x in ('spontaneous', 'passive')])
+                and not self.uiCheckAppend.isChecked()
+            ):
                 answer = QtWidgets.QMessageBox.question(
                     self,
                     'Is this a dud?',