+++ iblrig/test/tasks/test_passive_choice_world.py
@@ -6,6 +6,8 @@ from iblrig_tasks._iblrig_tasks_passiveChoiceWorld.task import Session as Passiv
 
 class TestInstantiatePassiveChoiceWorld(BaseTestCases.CommonTestInstantiateTask):
     def setUp(self) -> None:
-        self.task = PassiveChoiceWorldSession(**TASK_KWARGS)
+        session_id = 7
+        self.task = PassiveChoiceWorldSession(**TASK_KWARGS, session_template_id=session_id)
         self.task.mock()
+        assert np.unique(self.task.trials_table['session_id']) == [session_id]
         np.random.seed(12345)